# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Learning Platform API",
        "description": "API documentation for the Online Learning Platform",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/teacher/wallet/withdrawals/{id}": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Get withdrawal details",
                "operationId": "cc14c9c51abbb5e7cd8cf2b751fe952c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Withdrawal details"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Cancel a pending withdrawal",
                "operationId": "1988527c662ca305228d8b8934540fce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Withdrawal cancelled"
                    }
                }
            }
        },
        "/api/teacher/wallet/withdrawals": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "List withdrawal requests",
                "operationId": "3d4614dc101c7c26090f8b8a38f6e4b7",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of withdrawals"
                    }
                }
            }
        }
    },
...
}